Fix potential minor issue with gpiparser::speed
Hi. I run a similar project. I was looking for another other that had solved the problem I needed and while I struck out, I was looking at your code. I caught something that I'm not sure is wrong, ut sure looks like these are reversed. HTH.

When decoding contents of 8(30mps) and 9(40mps) it seems likely that one of these has a typo. Probably.

It's possible this is broken as I'm not building it or running anyhing beyond presubmit.. I'm not signing an CLA.